pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agent-intelligence-engine"
version = "2.0.0"
description = "Enterprise multi-agent intelligence engine for supply chain risk workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "google-genai",
  "langgraph",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "pydantic",
  "pydantic-settings",
  "psycopg[binary]",
  "redis",
  "structlog",
  "uvicorn",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[project.scripts]
maie = "maie.main:main"
maie-api = "maie.api.cli:main"
maie-eval = "maie.evaluation.cli:main"
maie-demo = "maie.demo.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"maie.demo" = ["web/*.html", "web/*.css", "web/*.js", "data/*.json"]

[tool.ruff]
line-length = 100
